requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-dotenv==1.0.0
supabase==2.0.3
httpx==0.24.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
groq==0.4.2
aiofiles==23.2.1
email-validator==2.1.0
python-multipart==0.0.6
